Add IPSReportList
Instead of storing all the reports in an array, create a IPSReportList
to handle managing all of the reports. Eventually this should subsume
the logic in compare.rb, and IPSReportList should be able to run the
comparison itself. The outside interface to Benchmark.ips should remain
unchanged, and the internal array list is returned. Eventually it might
be nice to return the IPSReportList instead. Creating the IPSReportList
means that other calculations on the list of reports can be stored in
this one location. I'm planning on taking advantage of this in a gem I'm
writing.