Compare against best result, not current result
When two pieces of code run the same number of iterations in the same
amount of time, that shouldn't be reported as one being "1.00x slower"
than the other. By comparing all slower results against the fastest
result, and not the other way around, we get some more realistic
reporting.