Feat: Support JSON output to STDOUT
I want to support `benchmark-ips` in
github-action-benchmark.
https://github.com/benchmark-action/github-action-benchmark

When it can write the result to STDOUT in JSON format, it would
be easy to support it.
This commits changes JSON output so that it can directly write
to a given argument if it responds to `write`.
It also adds tests and a section about JSON in README.