Report to stream (#111)
* Allow Stdout Report to go to any stream

It is a standard output report. Set it up to have any output

* MultiReport and remove noops

A multi report is a composite report

A noop report is slightly more efficient than a composite report with no outputs.
But it is good enough for our purposes

* converge suite and stdout calls

the handling of quiet is probably overkill.
Allowing the user to change their mind. Especially since it not per
test but for the whole quite run.

querying quite was fixed, but suite is now broken. For no suite it
used to return a noop (originally it returned nil) but now it returns
basically all output, including standard out. This is mostly because there
is no longer the distinction between a report and a suite. It will just
output of any kind.

Setting a benchmark to quiet and then adding a suite that is really a
standard output report directed to a specific output seems like a good use case.
This currently works, but quiet? will return false (when technically it is true)

This will work with existing benchmark suites, that do not have start_{warming,running} or footnote defined.
But will also call those methods if they are defined. This is a deviation from
the previous implementation.

Also, the previous case would replace the suite when assigned and this one will add more.
I would be difficult to add this detection (especially if someone chose to use a standard report
for the suite) and the use case doesn't seem as practical

previously suite returns a noop suite (when originally it would be nil)

* Rename StdoutReport to StreamReport

Now that this object will output to any stream, Stdout is not really
a valid name anymore. So renaming it to a stream report